==> 1_data_processing/01_process_protein_complexes.ipynb <==
## Generate map of entrez gene ids to protein complex(es)

**Input:** Protein complexes from CORUM.  

**Output:** Map of genes to protein complex membership

import pandas as pd
import numpy as np
import os

get_data_path = lambda folders, fname: os.path.normpath(os.environ['3RD_PARTY_DIR']+'/'+'/'.join(folders) +'/'+fname)
get_local_data_path = lambda folders, fname: os.path.normpath('../data/' +'/'.join(folders) +'/'+ fname)

file_CORUM_complexes = get_data_path(['CORUM'], 'allComplexes.txt')
file_complexes_map = get_local_data_path(['processed'], 'CORUM_entrez_map.csv')

all_complexes = pd.read_csv(file_CORUM_complexes, sep='\t')
all_complexes[:1]

# Reduce down to human complexes
human_complexes = all_complexes[all_complexes.Organism == 'Human']
human_complexes = human_complexes[['ComplexID', 'ComplexName', 'subunits(Entrez IDs)']]
print('Num complexes:', human_complexes.shape[0])
print('Max num sub-units:', human_complexes['subunits(Entrez IDs)'].apply(lambda x: x.count(';')).max() + 1)
print('Expected df length:', sum(human_complexes['subunits(Entrez IDs)'].apply(lambda x: x.count(';') + 1)))
human_complexes[:2]

# Note that some of the complexes include 'None' as one of the entrez ids, these should be filtered out, e.g.
display(human_complexes[human_complexes['ComplexID'] == 140])

# Flatten the list of entrez ids associated with each complex id
complex_map = pd.DataFrame(human_complexes['subunits(Entrez IDs)'].apply(lambda x: pd.Series(x.split(';'))).stack().reset_index(level=1, drop=True))
complex_map = complex_map.reset_index().rename(columns={'index':'complex_id', 0:'entrez_id'})

# Verify: num complexes, max num sub-units, expected length
assert(complex_map.complex_id.nunique() == human_complexes.shape[0])
assert(complex_map.complex_id.value_counts().max() == 
       human_complexes['subunits(Entrez IDs)'].apply(lambda x: x.count(';')).max() + 1)
assert(sum(human_complexes['subunits(Entrez IDs)'].apply(lambda x: x.count(';') + 1)) == complex_map.shape[0])

# Filter out entrez ids == None and fix entrez id type
complex_map = complex_map[complex_map.entrez_id != 'None']
complex_map = complex_map.astype({'entrez_id':'int'})

print('Map length:', complex_map.shape[0])
complex_map[:4]

complex_map.to_csv(file_complexes_map)
