==> 1_data_processing/04_process_depmap_expr.ipynb <==
## Filter DepMap Expression Data

Filter gene expression file down to cell lines that overlap with cell lines screened + only for paralogs

import pandas as pd
import os
import re

get_data_path = lambda folders, fname: os.path.normpath(os.environ['3RD_PARTY_DIR']+'/'+'/'.join(folders) +'/'+fname)
get_local_data_path = lambda folders, fname: os.path.normpath('../data/' +'/'.join(folders) +'/'+ fname)

# Input
file_gene_scores = get_local_data_path(['processed', 'depmap19Q1'], 'gene_scores_11_07_19.csv')
file_gene_expr = get_data_path(['depmap', '19Q1'], 'CCLE_depMap_19Q1_TPM.csv')
file_paralogs = get_local_data_path(['processed', 'ensembl93'], 'paralog_summary.csv')

# Output
file_filtered_expr = get_local_data_path(['processed', 'depmap19Q1'], 'filtered_gene_expression.csv')

all_expr_raw = pd.read_csv(file_gene_expr, index_col=0)

print(all_expr_raw.shape)
all_expr_raw[:1]

# Extract ensembl ids for column headings
get_gene_id = lambda x: re.search('[\w-]+\s\((\w+)\)', x).group(1)
all_expr = all_expr_raw.rename(columns=get_gene_id)
all_expr[:1]

all_scores = pd.read_csv(file_gene_scores, index_col=0).set_index('cell_line')

all_scores[:1]

paralogs = pd.read_csv(file_paralogs, index_col=0)
paralogs[:1]

# Drop non-overlapping cell lines (with CERES scores) and singletons
expr = all_expr.loc[all_expr.index.isin(all_scores.index), all_expr.columns.isin(paralogs.ensembl_id)]
print(all_expr.shape)
print(expr.shape)

expr.to_csv(file_filtered_expr)
